Reorganize time syscall into modules
