+++ src/lib/twizzler-abi/src/syscall/time/units.rs
@@ -1,0 +1,7 @@
+#[derive(Clone, Copy, Debug)]
+#[repr(transparent)]
+pub struct Seconds(pub u64);
+
+#[derive(Clone, Copy, Debug)]
+#[repr(transparent)]
+pub struct FemtoSeconds(pub u64);
